Working Alpha
This is an example of working application (.exe file). When you are
running this for the first time and click exit button it creates
SaveGame.ser in location where you have .exe file. So when you are
removing this app don't forget about removing this also. Thank you and
have fun.